# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API",
        "description": "API endpoints for Todo application",
        "version": "1.0.0"
    },
    "paths": {
        "/api/todos": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "List all todos",
                "operationId": "dd687d49d7f86d473ce56a7806776fc4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for todo title or details",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by todo status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "not-started",
                                "in-progress",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "created_at",
                                "status"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of todos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "todos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Todo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Todos"
                ],
                "summary": "Create a new todo",
                "operationId": "6f5da079c2dace66df52625276205a32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "details": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "default": "not-started",
                                        "enum": [
                                            "not-started",
                                            "in-progress",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "todo": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/todos/{todo}": {
            "get": {
                "tags": [
                    "Todos"
                ],
                "summary": "Get a specific todo",
                "operationId": "de852126e2655370ac90ca91baae44ea",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "todo": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Todos"
                ],
                "summary": "Update a todo",
                "operationId": "5318a346493eac3c7e3e48acba76cb5c",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "details": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "not-started",
                                            "in-progress",
                                            "completed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "todo": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Todos"
                ],
                "summary": "Delete a todo",
                "operationId": "6c5c891e96a4c8e36a743b73cbb71c7b",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "Todo ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "required": [
                    "title",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "details": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "not-started",
                            "in-progress",
                            "completed"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Todos",
            "description": "Todos"
        }
    ]
}